КилинаИВ_Добавила в контроллер работ студентов методы для добавления, именения и удаления катрочек работ. Добавила соотвествующие маршруты. Добавила в модель геттеры для получения http-путей к фотографиям и соотношения их сторон. Исправила баг попытки удаления фотографии не лежащей в пути /dep/

+++ app/Traits/Common.php
@@ -16,6 +16,6 @@ trait Common
             return $carry;
         });
 
-        Storage::delete($relPath);
+        $relPath && Storage::delete($relPath);
     }
 }
